+++ src/components/ItemCard/ItemCard.jsx
@@ -1,4 +1,4 @@
-import heartIcon from "../../../assets/images/ic_heart.png";
+import heartIcon from "../../assets/images/ic_heart.png";
 import "./ItemCard.css";
 
 function ItemCard({ item }) {
