+++ src/components/Pagination/PaginationBar.css
@@ -9,6 +9,7 @@
   height: 40px;
   border: 1px solid #e5e7eb;
   border-radius: 40px;
+  text-align: center;
   font-size: 16px;
   font-weight: 600;
   color: #6b7280;
