+++ src/pages/SamplePage/SamplePage.jsx
@@ -1,10 +1,10 @@
 import styles from "./SamplePage.module.css";
-import Icon from "../../components/Icon/Icon.jsx";
+import ItemCard from "../../components/ItemCard/ItemCard";
 
 const SamplePage = () => {
   return (
     <div className={styles.container}>
-      <Icon name="facebook" size={24} color="white" />
+      <ItemCard />
     </div>
   );
 };
