+++ src/pages/SamplePage/SamplePage.module.css
@@ -3,5 +3,4 @@
   flex-direction: column;
   gap: 2rem;
   padding: 15rem;
-  background-color: black;
 }
