--- src/styles/global.css
@@ -1,9 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-body {
-  font-family: "Pretendard", sans-serif;
-}
